refactor: extract expandGlobs to shared package and optimize file reads
- Create @graph-flow/shared package with expandGlobs utility
- Remove duplicated expandGlobs from graph and knowledge indexers
- Add optional content parameter to CodeParser.parse() to avoid double file reads
- Pass pre-read content from indexer to parser for efficiency
