+++ packages/graph/package.json
@@ -27,6 +27,7 @@
     "test": "bun test"
   },
   "dependencies": {
+    "@graph-flow/shared": "workspace:*",
     "@vue/compiler-sfc": "^3.5.0",
     "ts-morph": "^21.0.0"
   },
