+++ packages/shared/src/index.ts
@@ -1,0 +1,1 @@
+export { expandGlobs } from "./glob.ts";
